Renamed InstallCertificatesAndProfiles to installCodesigningFiles
