+++ autocodesign/write.go
@@ -1,0 +1,89 @@
+package autocodesign
+
+import (
+	"fmt"
+	"io/ioutil"
+	"os"
+	"path"
+
+	"github.com/bitrise-io/go-steputils/stepconf"
+	"github.com/bitrise-io/go-utils/log"
+	"github.com/bitrise-io/go-utils/pathutil"
+	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/appstoreconnect"
+	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/devportal"
+	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/keychain"
+)
+
+func installCodesigningFiles(codesignSettingsByDistributionType map[DistributionType]CodesignSettings, keychainPath string, keychainPassword stepconf.Secret) error {
+	fmt.Println()
+	log.Infof("Install certificates and profiles")
+
+	kc, err := keychain.New(keychainPath, keychainPassword)
+	if err != nil {
+		return fmt.Errorf("failed to initialize keychain: %s", err)
+	}
+
+	i := 0
+	for _, codesignSettings := range codesignSettingsByDistributionType {
+		log.Printf("certificate: %s", codesignSettings.Certificate.CommonName)
+
+		if err := kc.InstallCertificate(codesignSettings.Certificate, ""); err != nil {
+			return fmt.Errorf("failed to install certificate: %s", err)
+		}
+
+		log.Printf("profiles:")
+		for _, profile := range codesignSettings.ArchivableTargetProfilesByBundleID {
+			log.Printf("- %s", profile.Attributes().Name)
+
+			if err := writeProfile(profile); err != nil {
+				return fmt.Errorf("failed to write profile to file: %s", err)
+			}
+		}
+
+		for _, profile := range codesignSettings.UITestTargetProfilesByBundleID {
+			log.Printf("- %s", profile.Attributes().Name)
+
+			if err := writeProfile(profile); err != nil {
+				return fmt.Errorf("failed to write profile to file: %s", err)
+			}
+		}
+
+		if i < len(codesignSettingsByDistributionType)-1 {
+			fmt.Println()
+		}
+		i++
+	}
+
+	return nil
+}
+
+// writeProfile writes the provided profile under the `$HOME/Library/MobileDevice/Provisioning Profiles` directory.
+// Xcode uses profiles located in that directory.
+// The file extension depends on the profile's platform `IOS` => `.mobileprovision`, `MAC_OS` => `.provisionprofile`
+func writeProfile(profile devportal.Profile) error {
+	homeDir := os.Getenv("HOME")
+	profilesDir := path.Join(homeDir, "Library/MobileDevice/Provisioning Profiles")
+	if exists, err := pathutil.IsDirExists(profilesDir); err != nil {
+		return fmt.Errorf("failed to check directory (%s) for provisioning profiles: %s", profilesDir, err)
+	} else if !exists {
+		if err := os.MkdirAll(profilesDir, 0600); err != nil {
+			return fmt.Errorf("failed to generate directory (%s) for provisioning profiles: %s", profilesDir, err)
+		}
+	}
+
+	var ext string
+	switch profile.Attributes().Platform {
+	case appstoreconnect.IOS:
+		ext = ".mobileprovision"
+	case appstoreconnect.MacOS:
+		ext = ".provisionprofile"
+	default:
+		return fmt.Errorf("failed to write profile to file, unsupported platform: (%s). Supported platforms: %s, %s", profile.Attributes().Platform, appstoreconnect.IOS, appstoreconnect.MacOS)
+	}
+
+	name := path.Join(profilesDir, profile.Attributes().UUID+ext)
+	if err := ioutil.WriteFile(name, profile.Attributes().ProfileContent, 0600); err != nil {
+		return fmt.Errorf("failed to write profile to file: %s", err)
+	}
+	return nil
+}
